% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_x.R
\name{hello_x}
\alias{hello_x}
\title{Greet people}
\usage{
hello_x(x)
}
\arguments{
\item{x}{The name you enter}
}
\value{
The output will be printed from \code{print}
}
\description{
Greet people
}
\examples{
hello_x("Sam")
\dontrun{
hello_x("Bob")
}
}
